fix: useEffect 훅으로 분리
